Update docs for script for arbitrary variables
