+++ docs/index.md
@@ -137,6 +137,7 @@ INFO:ReX:area 0.000399, entropy difference 6.751189, insertion curve 0.964960, d
 :caption: Contents:
 background.md
 command_line.md
+scripts.md
 notebooks/intro
 config.md
 multiple.md
