+++ rex_xai/utils/_utils.py
@@ -11,7 +11,6 @@ from numpy.typing import NDArray
 from skimage.segmentation import mark_boundaries
 
 
-from rex_xai.input.input_data import Data
 from rex_xai.mutants.box import Box
 from rex_xai.responsibility.prediction import Prediction
 from rex_xai.utils.logger import logger
@@ -297,31 +296,6 @@ def set_boolean_mask_value(
     else:
         raise ReXError("mode not recognised")
 
-def validate_shape(data: Data, model_shape: Tuple[int, ...]) -> Data:
-    new_shape = list(model_shape)
-    logger.info(f"Validating model shape {new_shape} and making sure it matches the data's shape, which has a WIDTH of {data.model_width}, "
-                f"HEIGHT of {data.model_height},{"Depth of"+str(data.model_depth) if data.model_depth is not None else ""}")
-    for i, input_shape in enumerate(model_shape):
-        if input_shape == "W":
-             new_shape[i] = data.model_width
-        elif input_shape == "H":
-            new_shape[i] = data.model_height
-        elif input_shape == "D":
-            new_shape[i] = data.model_depth
-    # Make sure the data dimensions match the model shape
-    if data.model_order == "first":
-        assert data.model_width == new_shape[1]
-        assert data.model_height == new_shape[2]
-        if data.model_depth:
-            assert data.model_width == new_shape[3]
-    else:
-        assert data.model_width == new_shape[0]
-        assert data.model_height == new_shape[1]
-        if data.model_depth:
-            assert data.model_width == new_shape[2]
-
-    data.model_shape = tuple(new_shape)
-    return data
 
 def ff(obj, fmt):
     """
